some warnings fixed

+++ include/LightGBM/dataset.h
@@ -381,12 +381,12 @@ public:
   inline const std::vector<std::string>& feature_names() const { return feature_names_; }
 
   inline void set_feature_names(const std::vector<std::string>& feature_names) {
-    if (feature_names.size() != num_total_features_) {
+    if (feature_names.size() != static_cast<size_t>(num_total_features_)) {
       Log::Warning("size of feature_names error, should equal with total number of features");
       return;
     }
     feature_names_ = std::vector<std::string>(num_total_features_);
-    for (size_t i = 0; i < num_total_features_; ++i) {
+    for (int i = 0; i < num_total_features_; ++i) {
       feature_names_[i] = feature_names[i];
     }
   }
